Optimise on-eat backend code
Previously, to avoid the overlapping of snake's head and food in the view
phase there was an hardcoded position to spawn the food outside the board.
Now, the food will be added to the field in the same if-then-else expression
of the snake. In this way overlapping will not happen. This implied a code
reduction in "moveSnake" function and a return type slimming of "isChanged"
function.